backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy[asyncio]
psycopg2-binary 
pydantic
python-dotenv
alembic